+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# ver 20220409
+> - [试卷管理] 可以删除试卷
+> - [试卷管理] 为新增试卷/删除试卷添加了提示信息，返回操作成功或失败
+> - [数据库] 新增/删除试卷的数据库事务进行原子化
+> - [样式调整] 试卷管理/题库维护等相关界面样式调整
+> - [题库维护] 创建试卷以后自动清空创建题目的表格并刷新题库表格
+
 # ver 20220407
 > - [查看试卷] 查看试卷界面样式调整
 > - [查看试卷] 中选项自动生成A/B/C/D
