+++ TODO.md
@@ -81,7 +81,7 @@
         - [x] 题目和试卷可能需要多对多关系
     - [x] 题库维护界面
         - [x] 基本功能
-        - [ ] 界面样式调整
+        - [x] 界面样式调整
         - [ ] 题库管理多选功能
         - [ ] 题库管理筛选功能
         - [ ] 新增题目提交字段校验
@@ -96,13 +96,14 @@
         - [x] 展示题库，可以选择题目组装成试卷
         - [ ] 新增试卷界面可以直接添加题目
         - [x] 保存时被选中的题目引用次数+1
-        - [ ] 界面样式调整
+        - [x] 界面样式调整
         - [ ] 试卷有开放和结束时间(需调整数据库字段设计)，时间段以外禁止查看和作答试卷
         - [x] 试卷可以设置考试时间
         - [ ] 可以查看本课程的学生作答的试卷
         - [ ] 可以批改试卷
         - [ ] 可以设置考试类型(开卷/作业自动批改并即时返回成绩，闭卷/考试手动批改，完成后统一返回成绩)
-        - [ ] 可以删除试卷
+        - [x] 可以删除试卷
+            >   [模态框多次显示并点击确认以后会发送多个ajax请求](https://stackoverflow.com/questions/51402252/jquery-ajax-loaded-multiple-times-after-closing-the-modal)
         - [ ] 试题展示顺序和创建试卷时相同
         - [ ] 可以引用一张旧试卷
     - [ ] 课程学习测试与作业界面
@@ -115,3 +116,5 @@
 - [ ] **课程学习顶部课程名称/教师名称处，教师名称的链接可点击区域过长**
 - [ ] **课程图片裁剪限制最小尺寸**
 - [ ] **缓存前端插件到本地**
+- [ ] **数据库原子操作**
+    > [Rollback save with transaction atomic](https://stackoverflow.com/questions/34730385/django-rollback-save-with-transaction-atomic)
