+++ index/templates/index/ExaminationPages/questionBankManagement.html
@@ -2,6 +2,7 @@
 {% load static %}
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -18,29 +19,34 @@
     {% endblock extrastatic %}
 
 </head>
+
 <body>
     {% block content %}
     {{ block.super }}
     <div class="mainContainer">
         <div class="CreatorArea">
             <div class="Creator">
-                <button id="addRow">Add new row</button>
-                <button id="submit" type="button">Submit</button>
-                    <table id="CreatorTable" class="table table-striped table-bordered">
-                        <thead>
-                            <tr>
-                                <th></th>
-                                <th>题目内容</th>
-                                <th>题目类型</th>
-                                <th>题目分数</th>
-                                <th>是否公开</th>
-                                <th></th>
-                            </tr>
-                        </thead>
-                    </table>
+                <h3>新增题目</h3>
+                <div class="ToolBar">
+                    <div id="addRow"><span>新增题目</span></div>
+                    <div id="submit"><span>提交题库</span></div>
+                </div>
+                <table id="CreatorTable" class="table table-striped table-bordered">
+                    <thead>
+                        <tr>
+                            <th></th>
+                            <th>题目内容</th>
+                            <th>题目类型</th>
+                            <th>题目分数</th>
+                            <th>是否公开</th>
+                            <th></th>
+                        </tr>
+                    </thead>
+                </table>
             </div>
         </div>
         <div class="QuestionBankTable">
+            <h3>当前题库</h3>
             <table id="QuestionPreviewTable" class="table table-striped table-bordered">
                 <thead>
                     <tr>
@@ -64,7 +70,7 @@
     <!-- basic static -->
     <script src="{% static 'index/js/questionBank.js' %}"></script>
     {% endblock %}
-    <br/>
+    <br />
 </body>
-</html>
 
+</html>
